requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
python-dotenv
requests
flake8